django/requirements.txt
Django==1.7
MySQL-python==1.2.5
restless==2.0.1
requests==2.4.3
factory_boy==2.4.1
coverage==3.7.1
django-celery==3.1.16
redis==2.10.3